git push

+++ Model/DTO/Request/HoaDon/CreateHoaDonRequest.cs
@@ -12,7 +12,7 @@ namespace Models.DTO.Request.HoaDon
         public long BanId { get; set; }
         public Guid MaOrder { get; set; }
         public DateTime NgayTao { get; set; }
-        public double TongTien { get; set; }
+        public decimal TongTien { get; set; }
         public bool ThanhToan { get; set; } = false;
         public List<HoaDonSetMonAnRequest> Set { get; set; } = new List<HoaDonSetMonAnRequest>();
         public List<HoaDonMonAnRequest> MonAn { get; set; } = new List<HoaDonMonAnRequest>();
@@ -23,12 +23,12 @@ namespace Models.DTO.Request.HoaDon
     {
         public long MonAnId { get; set; }
         public int SoLuong { get; set; }
-        public double ThanhTien { get; set; }
+        public decimal ThanhTien { get; set; }
     }
     public class HoaDonSetMonAnRequest
     {
         public long SetId { get; set; }
         public int SoLuong { get; set; }
-        public double ThanhTien { get; set; }
+        public decimal ThanhTien { get; set; }
     }
 }
