+++ Model/Entities/HoaDon.cs
@@ -8,6 +8,6 @@ public partial class HoaDon
     public long Id { get; set; }    
     public Guid MaOrder {  get; set; }
     public DateTime NgayTao { get; set; } = DateTime.Now;
-    public double TongTien {  get; set; }
+    public decimal TongTien {  get; set; }
     public bool ThanhToan {  get; set; } = false;
 }
